signals is about to be done

+++ signals.c
@@ -6,35 +6,42 @@
 /*   By: kait-mar <kait-mar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/06 09:10:20 by kait-mar          #+#    #+#             */
-/*   Updated: 2021/01/12 12:14:59 by kait-mar         ###   ########.fr       */
+/*   Updated: 2021/01/19 15:47:58 by kait-mar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-int		inter_signal(int signum)
+void		inter_signal(int status)
 {
-	printf("\nCTR-C is pressed\n");
-	//print("my pid is %d\n", getpid());
-	//print("and my parent pid is %d\n", getppid());
-	if (g_process == 0)
+	int	pid;
+
+	//printf("\nCTR-C is pressed\n");
+	if ((pid = fork()) < 0)
+		printf("error in fork signal : %s\n", strerror(errno));
+	else if (pid == 0)
 	{
-		printf("before ignoring the g_process is %d\n", g_process);
-		printf("the signal is ignored\n");
-		//signal(SIGINT, SIG_IGN);
-		printf("after ignoring the g_process is %d\n", g_process);
-		g_process = 0;
-		return (0);
-	//	exit(EXIT_SUCCESS);
+		//printf("in Child\n");
+		if (g_process == 0)
+		{
+			//printf("the signal is ignored\n");
+			g_process = 0;
+		}
+		else
+		{
+			//printf("the signal is handled\n");
+			kill(g_process, SIGKILL);
+			g_process = 0;
+		}
+		exit(EXIT_SUCCESS);
 	}
 	else
 	{
-		printf("the signal is handled\n");
-		kill(g_process, SIGKILL);
-		g_process = 0;
-		return (1);
+		//printf("in parent\n");
+		if (waitpid(pid, &status, WUNTRACED) == -1)
+			printf("error in waipid signal : %s\n", strerror(errno));
 	}
-	printf("before returning\n");
+	prompt();
 }
 
 void	quit_signal(int signum)
@@ -43,12 +50,9 @@ void	quit_signal(int signum)
 	exit(EXIT_SUCCESS);
 }
 
-void	signal_handler(int signum)
+void	signal_handler(int *status)
 {
-	if (signal(SIGINT, inter_signal) == 0)
-	{
-		printf("before termination signal handler\n");
-		return ;
-	}
-	//signal(SIGQUIT, quit_signal);
+	signal(SIGINT, inter_signal);
+	signal(SIGQUIT, inter_signal);
+	//printf("before termination signal handler\n");
 }
